% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{full}
\alias{full}
\title{Full Keyboard}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 117 rows and 8 columns.
}
\usage{
full
}
\description{
A dataset containing rows and keys in a full mechanical keyboard.
}
\keyword{datasets}
