% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkeyboard.R
\name{ggkeyboard}
\alias{ggkeyboard}
\title{Plot a keyboard using ggplot2}
\usage{
ggkeyboard(
  keyboard = tkl,
  palette = keyboard_palette("pastel"),
  layout = c("ansi", "iso"),
  font_family = "Arial Unicode MS",
  font_size = 3,
  adjust_text_colour = TRUE,
  measurements = keyboard_measurements("default")
)
}
\arguments{
\item{keyboard}{Keyboard data. A data frame with the key name, what row of the keyboard it is in, and key width. Defaults to \code{\link{tkl}} (a tenkeyless layout). Other available keyboards are a full keyboard (\code{\link{full}}), 60\% keyboard (\code{\link{sixty_percent}}), and a basic mac keyboard (\code{\link{mac}}).}

\item{palette}{Colour palette. Defaults to \code{keyboard_palette("pastel")}. To use a custom palette, create a vector with the names described in \code{\link{keyboard_palette}}.}

\item{layout}{Keyboard layout - one of "ansi" or "iso". Defaults to "ansi".}

\item{font_family}{Font used. Defaults to "Arial Unicode MS". See the \code{extrafont} package for using fonts in ggplot2.}

\item{font_size}{Base font size. Defaults to 3.}

\item{adjust_text_colour}{Whether to lighten the text colour on dark keys. Defaults to TRUE.}

\item{measurements}{Measurements of various aspects of the keyboard key height and width, gaps between keys and rows, etc. See \code{\link{keyboard_measurements}}.}
}
\description{
Plot a keyboard using ggplot2
}
\examples{
\dontrun{
ggkeyboard()

ggkeyboard(sixty_percent, palette = keyboard_palette("cyberpunk"))
}
}
