% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_keys.R
\name{highlight_keys}
\alias{highlight_keys}
\title{Highlight keys}
\usage{
highlight_keys(ggkeyboard, keys, colour = "yellow", fill = NA, size = 1, ...)
}
\arguments{
\item{ggkeyboard}{An input keyboard from \code{\link{ggkeyboard}}.}

\item{keys}{Keys to highlight. Key names are available from the \code{key} column of the data passed to \code{ggkeyboard}.}

\item{colour}{Highlight outline colour.}

\item{fill}{Highlight fill colour.}

\item{size}{Highlight outline signs.}

\item{...}{Additional options, passed to \code{geom_rect}.}
}
\description{
Draw attention to keys in a \code{ggkeyboard}.
}
\examples{
\dontrun{
ggkeyboard(tkl) \%>\%
  highlight_keys(c("Alt Left", "Shift Left", "M"))
}
}
