% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{in_the_beginning_full}
\alias{in_the_beginning_full}
\title{In The Beginning Keyboard}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 95 rows and 8 columns.
}
\usage{
in_the_beginning_full
}
\description{
A dataset containing rows and keys in a full mechanical keyboard that spell out "IN THE BEGINNING WAS R MARKDOWN"
}
\keyword{datasets}
