% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyboard_measurements.R
\name{keyboard_measurements}
\alias{keyboard_measurements}
\title{Keyboard Measurements}
\usage{
keyboard_measurements(name = "default")
}
\arguments{
\item{name}{Measurement options name.}
}
\description{
Measurement options for \code{\link{ggkeyboard}}.
}
\details{
There are the following options:
\itemize{
\item key_height: Height of keys.
\item key_width: Base width of keys.
\item height_gap: Height gap between rows of keys.
\item width_gap: Width gap between keys in the same row.
\item segment_size: Size of segments used to draw arrows.
\item arrow_size; Size of arrow head.
}
}
