% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyboard_palette.R
\name{keyboard_palette}
\alias{keyboard_palette}
\title{Keyboard palettes}
\usage{
keyboard_palette(
  palette = c("pastel", "serika", "wahtsy", "cyberpunk", "magic", "varmilo")
)
}
\arguments{
\item{palette}{Name of palette.}
}
\description{
Built-in palettes for keyboards.
}
\details{
There are four palettes available:
\itemize{
\item "pastel" is just cute.
\item "serika" is based off the \href{https://drop.com/buy/drop-zambumon-mt3-serika-custom-keycap-set}{Drop + Zambumon MT3 Serika Custom Keycap Set}.
\item "wahtsy" is based off the \href{https://drop.com/buy/melgeek-mg-wahtsy-abs-doubleshot-keycap-set}{Melgeek MG Wahtsy ABS Doubleshot Keycap Set}.
\item "cyberpunk" is based off the \href{https://drop.com/buy/domikey-abs-doubleshot-sa-cyberpunk-pumper-keycap-set}{Domikey ABS Doubleshot SA Cyberpunk Pumper Keycap Set}.
\item "magic" is based off the \href{https://www.apple.com/shop/product/MLA22LL/A/magic-keyboard-us-english}{Apple magic keyboard}.
\item "varmilo" is based off the \href{https://drop.com/buy/varmilo-108-keyboard}{Varmilo VA108 Fullsize Keyboard}.
}

The palettes have the following fields:
\itemize{
\item background: Colour of background.
\item keyboard: Colour of keyboard.
\item alphanumeric: Colour of alpha-numeric keys and other common text keys (e.g. <, :, etc).
\item accent: Colour of accent keys (F1-4, F9-12, and the spacebar).
\item modifier: Colour of modifier keys (e.g. Shift, Print, Insert, etc).
\item numpad: Colour of numpad (non-modifier) keys (1-9).
\item arrow: Colour of arrow-pad keys.
\item light: Colour of lights on the keyboards.
\item text: Text colour.
}
}
\examples{
\dontrun{
library(scales)
show_col(keyboard_palette("pastel"))

ggkeyboard(palette = keyboard_palette("cyberpunk"))
}
}
