% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mac}
\alias{mac}
\title{Mac Keyboard}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 77 rows and 8 columns.
}
\usage{
mac
}
\description{
A dataset containing rows and keys in a mac keyboard, based off the Apple magic keyboard.
}
\keyword{datasets}
