% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sixty_percent}
\alias{sixty_percent}
\title{60\% Keyboard}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 61 rows and 8 columns.
}
\usage{
sixty_percent
}
\description{
A dataset containing rows and keys in a 60\% layout mechanical keyboard.
}
\keyword{datasets}
