% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{steno}
\alias{steno}
\title{Steno Keyboard}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 26 rows and 8 columns.
}
\usage{
steno
}
\description{
A dataset containing rows and keys in a stenography (steno) keyboard.
}
\keyword{datasets}
