% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tkl}
\alias{tkl}
\title{Tenkeyless}
\format{
An object of class \code{list} of length 6.
}
\usage{
tkl
}
\description{
A dataset containing rows and keys in a tenkeyless keyboard.
}
\keyword{datasets}
