% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tkl_dvorak}
\alias{tkl_dvorak}
\title{Tenkeyless Keyboard with dvorak layout}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 95 rows and 8 columns.
}
\usage{
tkl_dvorak
}
\description{
A dataset containing rows and keys in a tenkeyless keyboard with dvorak layout.
}
\keyword{datasets}
