% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tkl_iso}
\alias{tkl_iso}
\title{Tenkeyless ISO}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 97 rows and 6 columns.
}
\usage{
tkl_iso
}
\description{
A dataset containing rows and keys in a tenkeyless keyboard with an ISO layout.
}
\keyword{datasets}
