% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkeyboard.R
\name{ggkeyboard}
\alias{ggkeyboard}
\title{Plot a keyboard using ggplot2}
\usage{
ggkeyboard(
  data = tkl,
  key_height = 15/15.5,
  key_width = 1,
  height_gap = 2/15.5,
  width_gap = 2/15.5,
  font_size = 3,
  segment_size = 0.25,
  arrow_size = 0.03,
  font_family = "Avenir Next",
  keyboard_colour = "#fbbcb8",
  background_colour = "#fce9d0",
  alphanum_colour = "#bfdff6",
  accent_colour = "#a3e3c4",
  modifier_colour = "#78baeb",
  arrow_colour = "#c1b3ef",
  text_colour = "#5F5F5F",
  adjust_text_colour = TRUE,
  light_colour = clr_darken(keyboard_colour, 0.1),
  layout = c("ansi", "iso")
)
}
\arguments{
\item{data}{Keyboard data. A data frame with the key name, what row of the keyboard it is in, and key width. Defaults to \code{tkl}.}

\item{key_height}{Height of keys. Defaults to 15 / 15.5.}

\item{key_width}{Base width of keys. Defaults to 1.}

\item{height_gap}{Height gap between rows of keys. Defaults to 2 / 15.5.}

\item{width_gap}{Width gap between keys in the same row. Defaults to 2 / 15.5.}

\item{font_size}{Base font size. Defaults to 3.}

\item{segment_size}{Size of segments used to draw arrows. Defaults to 0.25.}

\item{arrow_size}{Size of arrow head. Defaults to 0.03 (npc).}

\item{font_family}{Font used. Defaults to "Avenir Next". See the \code{extrafont} package for using fonts.}

\item{keyboard_colour}{Colour of the keyboard. Defaults to "#fbbcb8".}

\item{background_colour}{Colour of the background. Defaults to "#fce9d0".}

\item{alphanum_colour}{Colour of alpha-numeric keys and other common text keys (e.g. <, :, etc). Defaults to "#bfdff6".}

\item{accent_colour}{Colour of accent keys (F1-4, F9-12, and the spacebar). Defaults to "#78baeb".}

\item{modifier_colour}{Colour of modifier keys (e.g. Shift, Print, Insert, etc). Defaults to "#78baeb".}

\item{arrow_colour}{Colour of arrow-pad keys. Defaults to "#c1b3ef".}

\item{text_colour}{Text colour. Defaults to "#5F5F5F".}

\item{adjust_text_colour}{Whether to ligthen the text colour on dark keys. Defaults to TRUE.}

\item{light_colour}{Colour of lights on the keyboards. Defaults to a 10\% darker version of \code{keyboard_colour}.}

\item{layout}{Keyboard layout - one of "ansi" or "iso". Defaults to "ansi".}
}
\description{
Plot a keyboard using ggplot2
}
\examples{
\dontrun{
ggkeyboard()

ggkeyboard(tkl, keyboard_colour = "#51504A", modifier_colour = "#ffce00",
accent_colour = "#454A49", alphanum_colour = "#EDEDD8", arrow_colour = "#454A49",
font_family = "Courier", background_colour = "lightgrey", light_colour = "#8aff2b")
}
}
