% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tkl}
\alias{tkl}
\title{Tenkeyless Keyboard}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 95 rows and 8 columns.
}
\usage{
tkl
}
\description{
A dataset containing rows and keys in a tenkeyless keyboard.
}
\keyword{datasets}
